requirements.txt
flask
flask-login
flask-sqlalchemy
passlib
waitress